app_folder/requirements.txt
Flask
boto3
python-dotenv
flask-session